Housekeeping RPC: support max datastream TTL at Realm level
Add a filed to realm cration, update and retrieval RPC to
specify the maximum TTL for datastreams.
Follow the ScyllaDB convention where no TTL is expressed by 0,
and allow the field to be missing in realm update (meaning
not modified).
Since 0 is also the zero value of an int64 field, a new realm
will not have a max datastream TTL unless explicitly set,
as it was before.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>